+++ frontend/app/src/entities/diff/ui/artifact-diff/artifact-content-diff.tsx
