+++ frontend/app/src/pages/branches/details.tsx
@@ -15,9 +15,9 @@ import { branchesState } from "@/entities/branches/stores";
 import { BranchDetails } from "@/entities/branches/ui/branch-details";
 import { BranchDefaultBadge } from "@/entities/branches/ui/branch-list-item/branch-default-badge";
 import { BranchStatusBadge } from "@/entities/branches/ui/branch-list-item/branch-status-badge";
-import { ArtifactsDiff } from "@/entities/diff/artifact-diff/artifacts-diff";
 import { FilesDiff } from "@/entities/diff/file-diff/files-diff";
 import { NodeDiff } from "@/entities/diff/node-diff";
+import { ArtifactsDiff } from "@/entities/diff/ui/artifact-diff/artifacts-diff";
 import { NodeMetadataPopover } from "@/entities/nodes/object/ui/object-details/node-metadata-popover";
 
 const BRANCH_TABS = {
